back-end/requirements.txt
# pip install -r requirements.txt

fastapi
uvicorn
torch
numpy
pandas
transformers
yfinance
python-dotenv
textblob
requests
stweet